swithced to  np.longdouble

+++ src/libSOGAtruncate.py
@@ -17,7 +17,7 @@ if hasattr(np, 'longdouble'):
     sogaType=np.float64
 else:
     print("np.float128 not available swithing to np.float64")
-    sogaType=np.float64
+    sogaType=np.float32
 
 def negate(trunc):
     """ Produces a string which is the logic negation of trunc """
